schema/users: realtime_sip_peers must contain 'port' column
When a peer is not found by username/password, asterisk does a
second lookup by host:port.

This commit FAKES the data, using the default port from the config,
which is most probably garbage. It should have used the registration
port IFF we were to trust that. Now, simply let the match fail,
just avoid the SQL error.